refactor: Field 컴포넌트를 layout 디렉토리 하위로 이동
